test(fixtures): split syntax audit report writers into a module
